requirements.txt
prokka		1.12
future		0.18.2
scikit-learn	0.23.2
biopython	1.78
numpy		1.19.2
pandas		1.1.3
xgboost	1.4.2